Break comparison header into its own component
Header in comparison table should act the same as it did before.
